Adding EventCards
These are just simple components which render an event as a floating card containing its state and ents state along with duration and properties. This introduced moment to the project which will be used a lot for time based operations and ReactTooltip for the condensed version. The condensed look produces a thinner, more vertical and reduced direct information view of the event suitable for embedding.