Bulk changes to text field: state management, labels, limits
This allows the state of a text field to managed by its parent and itself via listeners and by deriving its state from the props passed to it. Placeholders are now able to be provided which take priority over the name allowing better UI appearance. Min and max limits are now available, designed for number and date inputs but this is not enforced right now. Labels can also be omitted if they are being provided by the parent instead