Allowing select state to be managed by parent
Like text fields the select state is now optionally managed by the parent. There are also optional improvements to the key to prevent overlap in the rest of the project. Turns out this was pointless and I was just misreading the error but it does no harm so here we go