Creating a generic list page with auto generated entries
Similar to how some of the other components work, this one is passed a renderer and will just produce a list of cards. It also provides a default renderer which provides a best-effort attempt at rendering the components supporting IDs, colors and icons in a more native looking way. This can be provided or a better, more custom solution can be given.