Building a replacement for the normal calendar
This one is styled after the google calendar layout and supports the stacking of events. This will need some accessibility tweaks eventually and some scaling changes but it work for now. Additionally, I foolishly made this as a 5 day calendar instead of a 7 day one so will work on scaling this up next