Building the entire form system
I really should have done commits to structure this but tbh this was all built in about 4 sessions and was hastily thrown together. The system isn't that great but it works. I would like to go in and redo the styles at some point but the functionality is there. The code could very much do with some reformatting and some better consideration of state and props but it works. Will open a TODO for this in a moment